# task-manager/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Manager API",
    "description": "API for managing tasks and task-related entities",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API"
    },
    {
      "url": "/api/v1",
      "description": "Task Management API v1"
    }
  ],
...
  "paths": {
    "/api/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks",
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Task type or parent task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/task-types": {
      "get": {
        "tags": ["Task Types"],
        "summary": "Get all task types",
        "responses": {
          "200": {
            "description": "A list of task types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskType"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Task Types"],
        "summary": "Create a new task type",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskTypeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task type created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskType"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/task-templates": {
      "get": {
        "tags": ["Task Templates"],
        "summary": "Get all task templates",
        "responses": {
          "200": {
            "description": "A list of task templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTemplate"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Task Templates"],
        "summary": "Create a new task template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskTemplateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task template created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskTemplate"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/epics": {
      "get": {
        "tags": ["Epics"],
        "summary": "Get all epics",
        "responses": {
          "200": {
            "description": "List of epics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Epic"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Epics"],
        "summary": "Create a new epic",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpicInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Epic created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Epic"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jtbd": {
      "get": {
        "tags": ["Jobs To Be Done"],
        "summary": "Get all jobs to be done",
        "responses": {
          "200": {
            "description": "List of jobs to be done",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JTBD"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Jobs To Be Done"],
        "summary": "Create a new job to be done",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JTBDInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job to be done created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JTBD"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/stories": {
      "get": {
        "tags": ["User Stories"],
        "summary": "Get all user stories",
        "parameters": [
          {
            "name": "epicId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter stories by epic ID"
          },
          {
            "name": "jtbdId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter stories by job to be done ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user stories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserStory"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User Stories"],
        "summary": "Create a new user story",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStoryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User story created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStory"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Implement login feature"
          },
          "description": {
            "type": "string",
            "example": "Create login page with form validation"
          },
          "status": {
            "type": "string",
            "enum": ["TO_DO", "IN_PROGRESS", "IN_REVIEW", "DONE"],
            "example": "TO_DO"
          },
          "priority": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"],
            "example": "MEDIUM"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "example": "2023-12-31T23:59:59Z"
          },
          "estimatedHours": {
            "type": "number",
            "format": "float",
            "example": 8.5
          },
          "actualHours": {
            "type": "number",
            "format": "float",
            "example": 10.0
          },
          "assignedTo": {
            "type": "string",
            "example": "johndoe"
          },
          "createdBy": {
            "type": "string",
            "example": "janedoe"
          },
          "parentId": {
            "type": "integer",
            "nullable": true,
            "example": null
          },
          "taskTypeId": {
            "type": "integer",
            "example": 3
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "taskType": {
            "$ref": "#/components/schemas/TaskType"
          },
          "parent": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
      "TaskInput": {
        "type": "object",
        "required": ["title", "taskTypeId"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Implement login feature"
          },
          "description": {
            "type": "string",
            "example": "Create login page with form validation"
          },
          "taskTypeId": {
            "type": "integer",
            "example": 3
          },
          "status": {
            "type": "string",
            "enum": ["TO_DO", "IN_PROGRESS", "IN_REVIEW", "DONE"],
            "example": "TO_DO"
          },
          "priority": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"],
            "example": "MEDIUM"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "example": "2023-12-31T23:59:59Z"
          },
          "estimatedHours": {
            "type": "number",
            "format": "float",
            "example": 8.5
          },
          "actualHours": {
            "type": "number",
            "format": "float",
            "example": 10.0
          },
          "assignedTo": {
            "type": "string",
            "example": "johndoe"
          },
          "createdBy": {
            "type": "string",
            "example": "janedoe"
          },
          "parentId": {
            "type": "integer",
            "nullable": true,
            "example": null
          }
        }
      },
      "TaskType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Epic"
          },
          "description": {
            "type": "string",
            "example": "A large body of work that can be broken down into smaller stories"
          },
          "color": {
            "type": "string",
            "example": "#9b59b6"
          },
          "icon": {
            "type": "string",
            "example": "flag"
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TaskTypeInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Epic"
          },
          "description": {
            "type": "string",
            "example": "A large body of work that can be broken down into smaller stories"
          },
          "color": {
            "type": "string",
            "example": "#9b59b6"
          },
          "icon": {
            "type": "string",
            "example": "flag"
          },
          "isDefault": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "TaskTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Weekly Reporting Template"
          },
          "description": {
            "type": "string",
            "example": "Template for creating weekly reporting tasks"
          },
          "tasks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Prepare Financial Report"
                },
                "description": {
                  "type": "string",
                  "example": "Compile financial data for the week"
                },
                "taskTypeId": {
                  "type": "integer",
                  "example": 2
                },
                "dueDate": {
                  "type": "string",
                  "format": "date-time",
                  "example": null
                },
                "daysToComplete": {
                  "type": "integer",
                  "example": 2
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TaskTemplateInput": {
        "type": "object",
        "required": ["name", "tasks"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Weekly Reporting Template"
          },
          "description": {
            "type": "string",
            "example": "Template for creating weekly reporting tasks"
          },
          "tasks": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["title", "taskTypeId"],
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Prepare Financial Report"
                },
                "description": {
                  "type": "string",
                  "example": "Compile financial data for the week"
                },
                "taskTypeId": {
                  "type": "integer",
                  "example": 2
                },
                "dueDate": {
                  "type": "string",
                  "format": "date-time",
                  "example": null
                },
                "daysToComplete": {
                  "type": "integer",
                  "example": 2
                }
              }
            }
          }
        }
      },
...
      "Epic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "businessGoal": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "status": {
            "type": "string",
            "enum": ["planned", "in_progress", "completed", "blocked"]
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EpicInput": {
        "type": "object",
        "required": ["title", "description", "businessGoal"],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "businessGoal": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "status": {
            "type": "string",
            "enum": ["planned", "in_progress", "completed", "blocked"],
            "default": "planned"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "createdBy": {
            "type": "string"
          }
        }
      },
      "JTBD": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "epicId": {
            "type": "string"
          },
          "statement": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "contextualDetails": {
            "type": "string"
          },
          "motivations": {
            "type": "string"
          },
          "expectedOutcome": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "status": {
            "type": "string",
            "enum": ["identified", "validated", "in_progress", "completed"]
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JTBDInput": {
        "type": "object",
        "required": ["statement", "userType", "expectedOutcome"],
        "properties": {
          "epicId": {
            "type": "string"
          },
          "statement": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "contextualDetails": {
            "type": "string"
          },
          "motivations": {
            "type": "string"
          },
          "expectedOutcome": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "status": {
            "type": "string",
            "enum": ["identified", "validated", "in_progress", "completed"],
            "default": "identified"
          },
          "createdBy": {
            "type": "string"
          }
        }
      },
      "UserStory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "epicId": {
            "type": "string"
          },
          "jtbdId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "benefit": {
            "type": "string"
          },
          "acceptanceCriteria": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "points": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": ["backlog", "ready", "in_progress", "review", "done"]
          },
          "technicalConsiderations": {
            "type": "string"
          },
          "designConsiderations": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserStoryInput": {
        "type": "object",
        "required": ["userType", "action", "benefit"],
        "properties": {
          "epicId": {
            "type": "string"
          },
          "jtbdId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "benefit": {
            "type": "string"
          },
          "acceptanceCriteria": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "points": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": ["backlog", "ready", "in_progress", "review", "done"],
            "default": "backlog"
          },
          "technicalConsiderations": {
            "type": "string"
          },
          "designConsiderations": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          }
        }
      },
...
    }
  }
} 